refactor: make internal functions private
