+++ internal/stage11.go
@@ -18,7 +18,7 @@ func testCd2(stageHarness *test_case_harness.TestCaseHarness) error {
 		return err
 	}
 
-	directory, err := GetRandomDirectory()
+	directory, err := getRandomDirectory()
 	if err != nil {
 		return err
 	}
