+++ internal/stage_q2.go
@@ -16,7 +16,7 @@ func testQ2(stageHarness *test_case_harness.TestCaseHarness) error {
 	logger := stageHarness.Logger
 	shell := shell_executable.NewShellExecutable(stageHarness)
 
-	randomDir, err := GetShortRandomDirectory()
+	randomDir, err := getShortRandomDirectory()
 	if err != nil {
 		return err
 	}
